Refactor: 좋아요 취소 API에 409 추가
- 좋아요 API는 좋아요가 이미 눌린 상태였다면 409를 던지도록 되어 있었습니다.
- 좋아요 취소 API의 경우 이런 로직이 없는 상태여서 추가했습니다.
- 좋아요가 없는 상태이기 때문에 404를 할까 고민했지만 좋아요 상태 변경 요청이 현재 리소스의 상태와 충돌한다는 의미를 나타내기 때문에 Conflict가 더 적합하다고 판단했습니다.

+++ services/community/article.like.community.service.js
@@ -71,6 +71,18 @@ export const unlikeArticle = async ({
       throw new NotExistsError("게시글이 존재하지 않습니다"); // 404
     }
 
+    // 이미 좋아요가 눌렸는지 확인
+    const existingLike = await db.ArticleLikes.findOne({
+      where: {
+        articleId,
+        likedUserId,
+      },
+    });
+
+    if (!existingLike) {
+      // 좋아요가 눌리지 않은 경우
+      throw new AlreadyExistsError("좋아요가 존재하지 않습니다"); // 409
+    }
     // 좋아요 삭제
     await db.ArticleLikes.destroy({
       where: {
